SUMMARY: Convert project to recommended configuration based on OpenTAP's Python skeleton.
