+++ UR_Prototype.Api/UR_Prototype.Api.csproj
@@ -1,0 +1,30 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+    <PropertyGroup>
+        <TargetFramework>netstandard2.0</TargetFramework>
+        <OutputPath>../bin/</OutputPath>
+        <AppendTargetFrameworkToOutputPath>false</AppendTargetFrameworkToOutputPath>
+        <AppendRuntimeIdentifierToOutputPath>false</AppendRuntimeIdentifierToOutputPath>
+        <RootNamespace>Python.Test.CSharpApi</RootNamespace>
+    </PropertyGroup>
+    
+    <ItemGroup>
+        <PackageReference Include="OpenTAP" Version="9.19.4" />
+        <OpenTapPackageReference Include="Python" Version="3.0" />
+        <AdditionalOpenTapPackage Include="PythonExamples" Version="3.0" />
+        
+        <!-- Uncomment one of these lines to get a test plan editor installed when you build. -->
+        <OpenTapPackageReference Include="Editor CE"/>
+        <!-- <OpenTapPackageReference Include="EditorX"/> -->
+        <!-- <OpenTapPackageReference Include="TUI"/> -->
+    </ItemGroup>
+    
+    <ItemGroup>
+      <Reference Include="Microsoft.CSharp">
+        <HintPath>..\bin\Dependencies\Microsoft.CSharp.4.0.4.0\Microsoft.CSharp.dll</HintPath>
+      </Reference>
+    </ItemGroup>
+    <Target Name="AfterBuildLicensing" BeforeTargets="Build;PostBuildEvent">
+        <Exec Command=".\tap python search-path --add .. " WorkingDirectory="$(OutputPath)" />
+    </Target>
+</Project>
