+++ UR_Prototype.sln
@@ -1,0 +1,16 @@
+
+Microsoft Visual Studio Solution File, Format Version 12.00
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "UR_Prototype.CSharpApi", "UR_Prototype.Api\UR_Prototype.Api.csproj", "{752C3346-CE6E-4999-BD4F-71C203EEE4F7}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{752C3346-CE6E-4999-BD4F-71C203EEE4F7}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{752C3346-CE6E-4999-BD4F-71C203EEE4F7}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{752C3346-CE6E-4999-BD4F-71C203EEE4F7}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{752C3346-CE6E-4999-BD4F-71C203EEE4F7}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+EndGlobal
