+++ bin/.opentap_recent_logs
@@ -1,0 +1,4 @@
+C:\Users\shawn\OneDrive\Desktop\test\OpenTAP-Cobot-Plugin\bin\SessionLogs\SessionLog 2023-02-10 20-50-45.txt
+C:\Users\shawn\OneDrive\Desktop\test\OpenTAP-Cobot-Plugin\bin\SessionLogs\SessionLog 2023-02-10 20-51-38.txt
+C:\Users\shawn\OneDrive\Desktop\test\OpenTAP-Cobot-Plugin\bin\SessionLogs\SessionLog 2023-02-10 20-51-51.txt
+C:\Users\shawn\OneDrive\Desktop\test\OpenTAP-Cobot-Plugin\bin\SessionLogs\SessionLog 2023-02-10 20-52-18.txt
