+++ bin/Settings/Bench/Default/Connections.xml
@@ -1,0 +1,6 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<ConnectionSettings type="OpenTap.ConnectionSettings">
+  <Package.Dependencies>
+    <Package Name="OpenTAP" Version="^9.19.4+7959a308" />
+  </Package.Dependencies>
+</ConnectionSettings>
