+++ bin/Settings/Editor.xml
@@ -1,0 +1,25 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<EditorSettings type="Keysight.OpenTap.Wpf.EditorSettings">
+  <ShowWelcomeScreen>true</ShowWelcomeScreen>
+  <CheckForUpdates>false</CheckForUpdates>
+  <LogVisibility>ViewErrors, ViewWarnings, ViewMessages</LogVisibility>
+  <ColorTheme>Light</ColorTheme>
+  <ResultPopup>false</ResultPopup>
+  <ClearLog>true</ClearLog>
+  <ShowQuitConfirmation>true</ShowQuitConfirmation>
+  <ShowSaveBeforeExit>true</ShowSaveBeforeExit>
+  <ShowGuiLog>true</ShowGuiLog>
+  <EnsureUniqueStepName>true</EnsureUniqueStepName>
+  <ShowBrokenPackagesWindow>true</ShowBrokenPackagesWindow>
+  <EstimationWindowLength>10</EstimationWindowLength>
+  <LogPanelBufferSize>20000</LogPanelBufferSize>
+  <ShowCodeLinksInLog>false</ShowCodeLinksInLog>
+  <ShowPropertyInTooltip>false</ShowPropertyInTooltip>
+  <AutoAttachDebugger>false</AutoAttachDebugger>
+  <FocusMode>false</FocusMode>
+  <FocusModeDontAskMeAgain>false</FocusModeDontAskMeAgain>
+  <Package.Dependencies>
+    <Package Name="WPF Controls" Version="^9.19.1+26c29bbb" />
+    <Package Name="OpenTAP" Version="^9.19.4+7959a308" />
+  </Package.Dependencies>
+</EditorSettings>
