+++ bin/Settings/Engine.xml
@@ -1,0 +1,17 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<EngineSettings type="OpenTap.EngineSettings">
+  <SessionLogPath>SessionLogs/SessionLog &lt;Date&gt;.txt</SessionLogPath>
+  <PromptForMetaData>false</PromptForMetaData>
+  <AbortTestPlan>Step_Error</AbortTestPlan>
+  <OperatorName Metadata="Name">shawn</OperatorName>
+  <StationName Metadata="Station">DESKTOP-0AO2LQ5</StationName>
+  <ResultLatencyLimit>3</ResultLatencyLimit>
+  <LogTimestamper type="OpenTap.Diagnostic.AccurateStamper"></LogTimestamper>
+  <ResourceManagerType type="OpenTap.ResourceTaskManager">
+    <StaticResources />
+    <EnabledSteps />
+  </ResourceManagerType>
+  <Package.Dependencies>
+    <Package Name="OpenTAP" Version="^9.19.4+7959a308" />
+  </Package.Dependencies>
+</EngineSettings>
