+++ bin/Settings/GUI Panels.xml
@@ -1,0 +1,12 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<PanelSettings type="Keysight.OpenTap.Gui.PanelSettings">
+  <Layout>
+    <Base64>PD94bWwgdmVyc2lvbj0iMS4wIiBlbmNvZGluZz0idXRmLTgiPz4NCjxMYXlvdXRSb290Pg0KICA8Um9vdFBhbmVsIE9yaWVudGF0aW9uPSJWZXJ0aWNhbCI+DQogICAgPExheW91dFBhbmVsIE9yaWVudGF0aW9uPSJIb3Jpem9udGFsIiBEb2NrSGVpZ2h0PSIyLjQqIiBEb2NrTWluV2lkdGg9IjEwMCIgRG9ja01pbkhlaWdodD0iMTAwIj4NCiAgICAgIDxMYXlvdXRBbmNob3JhYmxlUGFuZSBEb2NrV2lkdGg9IjIuNyoiIERvY2tNaW5XaWR0aD0iMTAwIiBEb2NrTWluSGVpZ2h0PSIxMDAiPg0KICAgICAgICA8TGF5b3V0QW5jaG9yYWJsZSBBdXRvSGlkZU1pbldpZHRoPSIxMDAiIEF1dG9IaWRlTWluSGVpZ2h0PSIxMDAiIFRpdGxlPSJUZXN0IFBsYW4iIElzU2VsZWN0ZWQ9IlRydWUiIENvbnRlbnRJZD0iS2V5c2lnaHQuT3BlblRhcC5HdWkuVGVzdFBsYW5QbHVnaW4iIFRvb2xUaXA9IlRlc3QgUGxhbiBQYW5lbCIgTGFzdEFjdGl2YXRpb25UaW1lU3RhbXA9IjAyLzEwLzIwMjMgMTk6Mzk6NTEiIC8+DQogICAgICA8L0xheW91dEFuY2hvcmFibGVQYW5lPg0KICAgICAgPExheW91dEFuY2hvcmFibGVQYW5lIERvY2tNaW5XaWR0aD0iMTAwIiBEb2NrTWluSGVpZ2h0PSIxMDAiPg0KICAgICAgICA8TGF5b3V0QW5jaG9yYWJsZSBBdXRvSGlkZU1pbldpZHRoPSIxMDAiIEF1dG9IaWRlTWluSGVpZ2h0PSIxMDAiIFRpdGxlPSJUZXN0IFN0ZXAgU2V0dGluZ3MiIElzU2VsZWN0ZWQ9IlRydWUiIENvbnRlbnRJZD0iS2V5c2lnaHQuT3BlblRhcC5HdWkuU3RlcFNldHRpbmdzUGx1Z2luIiBUb29sVGlwPSJTdGVwIFNldHRpbmdzIFBhbmVsIiAvPg0KICAgICAgPC9MYXlvdXRBbmNob3JhYmxlUGFuZT4NCiAgICA8L0xheW91dFBhbmVsPg0KICAgIDxMYXlvdXRBbmNob3JhYmxlUGFuZSBEb2NrTWluV2lkdGg9IjEwMCIgRG9ja01pbkhlaWdodD0iMTAwIj4NCiAgICAgIDxMYXlvdXRBbmNob3JhYmxlIEF1dG9IaWRlTWluV2lkdGg9IjEwMCIgQXV0b0hpZGVNaW5IZWlnaHQ9IjEwMCIgVGl0bGU9IkxvZyIgSXNTZWxlY3RlZD0iVHJ1ZSIgQ29udGVudElkPSJLZXlzaWdodC5PcGVuVGFwLkd1aS5Mb2dQYW5lbFBsdWdpbiIgVG9vbFRpcD0iTG9nIFBhbmVsIiAvPg0KICAgIDwvTGF5b3V0QW5jaG9yYWJsZVBhbmU+DQogIDwvUm9vdFBhbmVsPg0KICA8VG9wU2lkZSAvPg0KICA8UmlnaHRTaWRlIC8+DQogIDxMZWZ0U2lkZSAvPg0KICA8Qm90dG9tU2lkZSAvPg0KICA8RmxvYXRpbmdXaW5kb3dzIC8+DQogIDxIaWRkZW4gLz4NCjwvTGF5b3V0Um9vdD4=</Base64>
+  </Layout>
+  <LayoutXml />
+  <VisiblePanels />
+  <Package.Dependencies>
+    <Package Name="Editor" Version="^9.19.1+26c29bbb" />
+    <Package Name="OpenTAP" Version="^9.19.4+7959a308" />
+  </Package.Dependencies>
+</PanelSettings>
