+++ bin/UR_Prototype.Api.deps.json
@@ -1,0 +1,91 @@
+{
+  "runtimeTarget": {
+    "name": ".NETStandard,Version=v2.0/",
+    "signature": ""
+  },
+  "compilationOptions": {},
+  "targets": {
+    ".NETStandard,Version=v2.0": {},
+    ".NETStandard,Version=v2.0/": {
+      "UR_Prototype.Api/1.0.0": {
+        "dependencies": {
+          "NETStandard.Library": "2.0.3",
+          "OpenTAP": "9.19.4",
+          "Microsoft.CSharp": "4.0.4.0",
+          "OpenTap.Python": "3.0.1.0"
+        },
+        "runtime": {
+          "UR_Prototype.Api.dll": {}
+        }
+      },
+      "Microsoft.NETCore.Platforms/1.1.0": {},
+      "NETStandard.Library/2.0.3": {
+        "dependencies": {
+          "Microsoft.NETCore.Platforms": "1.1.0"
+        }
+      },
+      "OpenTAP/9.19.4": {
+        "runtime": {
+          "lib/netstandard2.0/OpenTap.dll": {
+            "assemblyVersion": "9.19.4.0",
+            "fileVersion": "9.19.4.0"
+          }
+        }
+      },
+      "Microsoft.CSharp/4.0.4.0": {
+        "runtime": {
+          "Microsoft.CSharp.dll": {
+            "assemblyVersion": "4.0.4.0",
+            "fileVersion": "4.6.26515.6"
+          }
+        }
+      },
+      "OpenTap.Python/3.0.1.0": {
+        "runtime": {
+          "OpenTap.Python.dll": {
+            "assemblyVersion": "3.0.1.0",
+            "fileVersion": "1.0.0.0"
+          }
+        }
+      }
+    }
+  },
+  "libraries": {
+    "UR_Prototype.Api/1.0.0": {
+      "type": "project",
+      "serviceable": false,
+      "sha512": ""
+    },
+    "Microsoft.NETCore.Platforms/1.1.0": {
+      "type": "package",
+      "serviceable": true,
+      "sha512": "sha512-kz0PEW2lhqygehI/d6XsPCQzD7ff7gUJaVGPVETX611eadGsA3A877GdSlU0LRVMCTH/+P3o2iDTak+S08V2+A==",
+      "path": "microsoft.netcore.platforms/1.1.0",
+      "hashPath": "microsoft.netcore.platforms.1.1.0.nupkg.sha512"
+    },
+    "NETStandard.Library/2.0.3": {
+      "type": "package",
+      "serviceable": true,
+      "sha512": "sha512-st47PosZSHrjECdjeIzZQbzivYBJFv6P2nv4cj2ypdI204DO+vZ7l5raGMiX4eXMJ53RfOIg+/s4DHVZ54Nu2A==",
+      "path": "netstandard.library/2.0.3",
+      "hashPath": "netstandard.library.2.0.3.nupkg.sha512"
+    },
+    "OpenTAP/9.19.4": {
+      "type": "package",
+      "serviceable": true,
+      "sha512": "sha512-+pyhYs2k6dJy45bEIr+0yyKEBWA47xByneO9gywRWma1syJnuhPJs9MglasXGL+UdeWVvBpnRRS1/3YVCSw5/Q==",
+      "path": "opentap/9.19.4",
+      "hashPath": "opentap.9.19.4.nupkg.sha512"
+    },
+    "Microsoft.CSharp/4.0.4.0": {
+      "type": "reference",
+      "serviceable": false,
+      "sha512": ""
+    },
+    "OpenTap.Python/3.0.1.0": {
+      "type": "reference",
+      "serviceable": false,
+      "sha512": ""
+    }
+  }
+}
