+++ bin/~last_session.21748.RecoveryTapPlan
@@ -1,0 +1,21 @@
+<?xml version="1.0" encoding="utf-8"?>
+<RecoveryFile type="Keysight.OpenTap.Gui.MainWindow+RecoveryFile">
+  <FilePath>Untitled</FilePath>
+  <TestPlan type="OpenTap.TestPlan">
+    <Steps>
+      <TestStep type="UR_Prototype.MoveCobot.MoveCobot" Id="40ed29ed-4ba7-45a3-ae22-f7c6bec176d5">
+        <UR3e_cobot Source="OpenTap.InstrumentSettings">UR3e</UR3e_cobot>
+        <Command>movej([0, 0, 0, 0, 0, 0], a=1.2, v=1.05)</Command>
+        <Name Metadata="Step Name">Move Cobot</Name>
+      </TestStep>
+      <TestStep type="UR_Prototype.TestStepExample.TestStepExample" Id="e4b184fd-ed64-480f-9fca-632ff8cad8cf">
+        <Number>1000</Number>
+        <Name Metadata="Step Name">Test Step Example</Name>
+      </TestStep>
+    </Steps>
+  </TestPlan>
+  <Package.Dependencies>
+    <Package Name="Editor" Version="^9.19.1+26c29bbb" />
+    <Package Name="OpenTAP" Version="^9.19.4+7959a308" />
+  </Package.Dependencies>
+</RecoveryFile>
