+++ package.xml
@@ -1,0 +1,20 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<Package Name="UR_Prototype" xmlns="http://keysight.com/Schemas/tap" Version="$(GitVersion)" OS="Windows,Linux,MacOS">
+    <Description>
+        UR_Prototype
+        <Prerequisites> Python (>3.7) </Prerequisites>
+    </Description>
+    
+    <Files>
+        <File Path="UR_Prototype/*.py" >
+            <ProjectFile/>
+        </File>
+                    
+        <File Path="Packages/UR_Prototype/UR_Prototype.Api.dll" SourcePath="bin/UR_Prototype.Api.dll" >
+            <ProjectFile/>
+        </File>
+        <File Path="Readme.md" >
+            <ProjectFile/>
+        </File>
+    </Files>
+</Package>
